refactor: Modularization for Chat, core file js
- Add chat.ui.js for ui
- Add ./chat/index.js for import chat.js & chat.ui.js
- Move keyPair.js & websocket to /js/core/
