+++ public/js/core/keyPair.js
