+++ public/js/load_animation.js
@@ -1,5 +1,3 @@
-const link = document.querySelector('#new-fish');
-
 const loadComplete = () => {
     const loader = document.querySelector('#loader');
     const width = window.innerWidth;
@@ -12,41 +10,4 @@ const loadComplete = () => {
     setTimeout(() => { loader.remove() }, 700);
 }
 
-function makeEditable() {
-    const currentSpan = document.querySelector('#span-new-fish');
-    if (!currentSpan) return;
-
-    const input = document.createElement('input');
-    input.type = 'text';
-    input.value = '';
-    input.className = 'bg-gray-800 text-white px-2 py-1 rounded w-full focus:outline-none';
-    input.id = 'input-new-fish';
-
-    const save = () => {
-        const newSpan = document.createElement('span');
-        newSpan.id = 'span-new-fish';
-        newSpan.className = 'px-2 py-1 w-full';
-        newSpan.textContent = "New Fish Basket";
-        link.replaceChild(newSpan, input);
-    };
-
-    input.addEventListener('keydown', function (e) {
-        if (e.key === 'Enter') {
-            save();
-        }
-    });
-
-    input.addEventListener('blur', () => {
-        setTimeout(() => {
-            if (link.contains(input)) {
-                save();
-            }
-        }, 0);
-        //When entering the input, it will call the blur event immediately, so we need to delay it a bit
-    });
-    link.replaceChild(input, currentSpan);
-    input.focus();
-}
-
 loadComplete();
-link.addEventListener('click', makeEditable);
