+++ public/js/m/chat/chat.ui.js
@@ -1,0 +1,38 @@
+import { startPM } from "./chat.js";
+
+export function newFishBasket(newBtn) {
+    const currentSpan = document.querySelector('#span-new-fish');
+    if (!currentSpan) return;
+
+    const input = document.createElement('input');
+    input.type = 'text';
+    input.value = '';
+    input.className = 'bg-gray-800 text-white px-2 py-1 rounded w-full focus:outline-none';
+    input.id = 'input-new-fish';
+
+    const save = () => {
+        const newSpan = document.createElement('span');
+        newSpan.id = 'span-new-fish';
+        newSpan.className = 'px-2 py-1 w-full';
+        newSpan.textContent = "New Fish Basket";
+        newBtn.replaceChild(newSpan, input);
+    };
+
+    input.addEventListener('keydown', function (e) {
+        if (e.key === 'Enter') {
+            startPM(input.value);
+            save();
+        }
+    });
+
+    input.addEventListener('blur', () => {
+        setTimeout(() => {
+            if (newBtn.contains(input)) {
+                save();
+            }
+        }, 0);
+        //When entering the input, it will call the blur event immediately, so we need to delay it a bit
+    });
+    newBtn.replaceChild(input, currentSpan);
+    input.focus();
+}
